+++ packages/adapters/eth/src/websocket.rs
@@ -1,5 +1,5 @@
 use std::{
-    cmp::min, num::NonZeroU32, ops::RangeInclusive, str::FromStr, sync::Arc, time::Duration,
+    cmp::min, num::NonZeroU32, ops::RangeInclusive,
 };
 
 use crate::{
@@ -8,24 +8,19 @@ use crate::{
 };
 use ::metrics::RegistersMetrics;
 use alloy::{
-    consensus::{SignableTransaction, Transaction},
-    eips::{
-        BlockNumberOrTag,
-        eip4844::DATA_GAS_PER_BLOB,
-    },
+    consensus::Transaction,
+    eips::{BlockNumberOrTag, eip4844::DATA_GAS_PER_BLOB},
     network::{Ethereum, EthereumWallet, TransactionBuilder, TransactionBuilder4844, TxSigner},
-    primitives::{Address, B256, ChainId},
+    primitives::Address,
     providers::{
         Provider, ProviderBuilder, SendableTx, WsConnect,
         utils::{EIP1559_FEE_ESTIMATION_PAST_BLOCKS, Eip1559Estimation},
     },
     pubsub::PubSubFrontend,
     rpc::types::{FeeHistory, TransactionReceipt, TransactionRequest},
-    signers::{Signature, local::PrivateKeySigner},
     sol,
 };
 use itertools::Itertools;
-use serde::Deserialize;
 use services::{
     state_committer::port::l1::Priority,
     types::{
@@ -35,11 +30,9 @@ use services::{
 use tracing::info;
 use url::Url;
 
-use crate::{
-    AwsClient, AwsConfig, Error, Result, blob_encoder,
-    provider::L1Provider,
-};
+use crate::{Error, Result, blob_encoder, provider::L1Provider};
 
+pub mod config;
 pub mod factory;
 mod metrics;
 
@@ -88,15 +81,15 @@ pub struct WebsocketClient {
     contract: FuelStateContract,
     commit_interval: NonZeroU32,
     metrics: Metrics,
-    tx_config: TxConfig,
+    tx_config: config::TxConfig,
 }
 
 impl WebsocketClient {
     pub async fn connect(
         url: Url,
         contract_address: Address,
-        signers: Signers,
-        tx_config: TxConfig,
+        signers: config::Signers,
+        tx_config: config::TxConfig,
     ) -> Result<Self> {
         let blob_poster_address = signers
             .blob
@@ -454,219 +447,12 @@ impl L1Provider for WebsocketClient {
     }
 }
 
-#[derive(Debug, Clone, PartialEq)]
-pub enum L1Key {
-    Kms(String),
-    Private(String),
-}
-
-impl<'a> serde::Deserialize<'a> for L1Key {
-    fn deserialize<D>(deserializer: D) -> std::result::Result<Self, D::Error>
-    where
-        D: serde::Deserializer<'a>,
-    {
-        let value = String::deserialize(deserializer)?;
-        if let Some(k) = value.strip_prefix("Kms(").and_then(|s| s.strip_suffix(')')) {
-            Ok(L1Key::Kms(k.to_string()))
-        } else if let Some(k) = value
-            .strip_prefix("Private(")
-            .and_then(|s| s.strip_suffix(')'))
-        {
-            Ok(L1Key::Private(k.to_string()))
-        } else {
-            Err(serde::de::Error::custom("invalid L1Key format"))
-        }
-    }
-}
-
-#[derive(Debug, Clone, Deserialize, PartialEq)]
-pub struct L1Keys {
-    /// The eth key authorized by the L1 bridging contracts to post block commitments.
-    pub main: L1Key,
-    /// The eth key for posting L2 state to L1.
-    pub blob: Option<L1Key>,
-}
-
-impl L1Keys {
-    pub fn uses_aws(&self) -> bool {
-        matches!(self.main, L1Key::Kms(_)) || matches!(self.blob, Some(L1Key::Kms(_)))
-    }
-}
-
-#[derive(Debug, Clone)]
-pub struct TxConfig {
-    pub tx_max_fee: u128,
-    pub send_tx_request_timeout: Duration,
-    pub acceptable_priority_fee_percentage: AcceptablePriorityFeePercentages,
-}
-
-#[cfg(feature = "test-helpers")]
-impl Default for TxConfig {
-    fn default() -> Self {
-        Self {
-            tx_max_fee: u128::MAX,
-            send_tx_request_timeout: Duration::from_secs(10),
-            acceptable_priority_fee_percentage: AcceptablePriorityFeePercentages::default(),
-        }
-    }
-}
-
-#[derive(Debug, Clone, Copy)]
-pub struct AcceptablePriorityFeePercentages {
-    min: f64,
-    max: f64,
-}
-
-#[cfg(feature = "test-helpers")]
-impl Default for AcceptablePriorityFeePercentages {
-    fn default() -> Self {
-        Self::new(20., 20.).expect("valid reward percentile range")
-    }
-}
-
-impl AcceptablePriorityFeePercentages {
-    pub fn new(min: f64, max: f64) -> Result<Self> {
-        if min > max {
-            return Err(crate::Error::Other(
-                "min reward percentile must be less than or equal to max reward percentile"
-                    .to_string(),
-            ));
-        }
-
-        if min <= 0.0 || max > 100.0 {
-            return Err(crate::Error::Other(
-                "reward percentiles must be > 0 and <= 100".to_string(),
-            ));
-        }
-
-        Ok(Self { min, max })
-    }
-
-    pub fn apply(&self, priority: Priority) -> f64 {
-        let min = self.min;
-
-        let increase = (self.max - min) * priority.get() / 100.;
-
-        (min + increase).min(self.max)
-    }
-}
-
-// This trait is needed because you cannot write `dyn TraitA + TraitB` except when TraitB is an
-// auto-trait.
-trait CompositeSigner: alloy::signers::Signer + TxSigner<Signature> {}
-impl<T: alloy::signers::Signer + TxSigner<Signature>> CompositeSigner for T {}
-
-#[derive(Clone)]
-pub struct Signer {
-    signer: Arc<dyn CompositeSigner + 'static + Send + Sync>,
-    chain_id: Option<ChainId>,
-}
-
-#[async_trait::async_trait]
-impl TxSigner<Signature> for Signer {
-    fn address(&self) -> Address {
-        TxSigner::<Signature>::address(&self.signer)
-    }
-
-    async fn sign_transaction(
-        &self,
-        tx: &mut dyn SignableTransaction<Signature>,
-    ) -> alloy::signers::Result<Signature> {
-        TxSigner::<Signature>::sign_transaction(&self.signer, tx).await
-    }
-}
-
-#[async_trait::async_trait]
-impl alloy::signers::Signer<Signature> for Signer {
-    async fn sign_hash(&self, hash: &B256) -> alloy::signers::Result<Signature> {
-        self.signer.sign_hash(hash).await
-    }
-
-    fn address(&self) -> Address {
-        alloy::signers::Signer::<Signature>::address(&*self.signer)
-    }
-
-    fn chain_id(&self) -> Option<ChainId> {
-        self.chain_id
-    }
-
-    fn set_chain_id(&mut self, chain_id: Option<ChainId>) {
-        self.chain_id = chain_id;
-    }
-}
-
-impl Signer {
-    pub async fn make_aws_signer(client: &AwsClient, key: String) -> Result<Self> {
-        let signer = client.make_signer(key).await?;
-        let chain_id = alloy::signers::Signer::chain_id(&signer);
-
-        Ok(Signer {
-            signer: Arc::new(signer),
-            chain_id,
-        })
-    }
-
-    pub fn make_private_key_signer(key: &str) -> Result<Self> {
-        let signer = PrivateKeySigner::from_str(key)
-            .map_err(|_| Error::Other("Invalid private key".to_string()))?;
-        let chain_id = signer.chain_id();
-
-        Ok(Signer {
-            signer: Arc::new(signer),
-            chain_id,
-        })
-    }
-}
-
-pub struct Signers {
-    pub main: Signer,
-    pub blob: Option<Signer>,
-}
-
-impl Clone for Signers {
-    fn clone(&self) -> Self {
-        Self {
-            main: self.main.clone(),
-            blob: self.blob.clone(),
-        }
-    }
-}
-
-impl Signers {
-    pub async fn for_keys(keys: L1Keys) -> Result<Self> {
-        let aws_client = if keys.uses_aws() {
-            let config = AwsConfig::from_env().await;
-            Some(AwsClient::new(config))
-        } else {
-            None
-        };
-
-        let blob_signer = match keys.blob {
-            Some(L1Key::Kms(key)) => {
-                Some(Signer::make_aws_signer(aws_client.as_ref().expect("is set"), key).await?)
-            }
-            Some(L1Key::Private(key)) => Some(Signer::make_private_key_signer(&key)?),
-            None => None,
-        };
-
-        let main_signer = match keys.main {
-            L1Key::Kms(key) => Signer::make_aws_signer(&aws_client.expect("is set"), key).await?,
-            L1Key::Private(key) => Signer::make_private_key_signer(&key)?,
-        };
-
-        Ok(Self {
-            main: main_signer,
-            blob: blob_signer,
-        })
-    }
-}
-
 #[cfg(test)]
 mod tests {
     use pretty_assertions::assert_eq;
     use services::state_committer::port::l1::Priority;
 
-    use super::L1Key;
+    use super::config::L1Key;
 
     #[test]
     fn can_deserialize_private_key() {
@@ -674,7 +460,7 @@ mod tests {
         let val = r#""Private(0x1234)""#;
 
         // when
-        let key: L1Key = serde_json::from_str(val).unwrap();
+        let key: config::L1Key = serde_json::from_str(val).unwrap();
 
         // then
         assert_eq!(key, L1Key::Private("0x1234".to_owned()));
@@ -686,7 +472,7 @@ mod tests {
         let val = r#""Kms(0x1234)""#;
 
         // when
-        let key: L1Key = serde_json::from_str(val).unwrap();
+        let key: config::L1Key = serde_json::from_str(val).unwrap();
 
         // then
         assert_eq!(key, L1Key::Kms("0x1234".to_owned()));
@@ -695,7 +481,7 @@ mod tests {
     #[test]
     fn lowest_priority_gives_min_priority_fee_perc() {
         // given
-        let sut = super::AcceptablePriorityFeePercentages::new(20., 40.).unwrap();
+        let sut = super::config::AcceptablePriorityFeePercentages::new(20., 40.).unwrap();
 
         // when
         let fee_perc = sut.apply(Priority::MIN);
@@ -707,7 +493,7 @@ mod tests {
     #[test]
     fn medium_priority_gives_middle_priority_fee_perc() {
         // given
-        let sut = super::AcceptablePriorityFeePercentages::new(20., 40.).unwrap();
+        let sut = super::config::AcceptablePriorityFeePercentages::new(20., 40.).unwrap();
 
         // when
         let fee_perc = sut.apply(Priority::new(50.).unwrap());
@@ -719,7 +505,7 @@ mod tests {
     #[test]
     fn highest_priority_gives_max_priority_fee_perc() {
         // given
-        let sut = super::AcceptablePriorityFeePercentages::new(20., 40.).unwrap();
+        let sut = super::config::AcceptablePriorityFeePercentages::new(20., 40.).unwrap();
 
         // when
         let fee_perc = sut.apply(Priority::MAX);
@@ -782,7 +568,7 @@ mod tests {
                 provider.clone(),
             ),
             commit_interval: 3.try_into().unwrap(),
-            tx_config: TxConfig::default(),
+            tx_config: config::TxConfig::default(),
             metrics: Default::default(),
         };
 
@@ -861,7 +647,7 @@ mod tests {
                 provider.clone(),
             ),
             commit_interval: 3.try_into().unwrap(),
-            tx_config: TxConfig {
+            tx_config: config::TxConfig {
                 tx_max_fee,
                 ..Default::default()
             },
